Deleted License file
Deleted the License file as the license is not applied to each individual file yet. Also, the placeholders were not yet filled in so this project is not yet licensed and this file can be deleted. I may license this project under GPL later, once I am further in the development and writing the javadoc for this project. Until then this project is protected under german copyright laws.