Finalize beta for 2.0.0
Putting all the pieces together to allow LANChat to enter beta testing for 2.0.0.

Some features from 1.x.x are still missing, notably Plugins and Archiving, but both are not required for now.

With this change the update mechanism was also overhauled, so that we no longer need a centrally maintained update infrastructure. Instead, clients can provide download links to updates.